created search user api, authorisation middleware, defined chat api routes, and created one on one chat api
Created search user api to search user from DB. Returns users related to search except the currently logged in user. Created authorisation middleware to determine if the user is logged in or not. Uses JWT token to authorise user. Defined the chat API routes. Created the access chat route. We check if the users are registered in the DB, if they are we check if chat with the user exists in the chat model and db and if isGroup chat is false, if such case it does exist, then we return that chat. If it doesnt , we create a chat with that user, set isGroup chat to false and store it in our db. we then fetch this  chat we created from the db using the chat id and send it as response to the request.